+++ src/main/java/org/logistic/model/Point.java
@@ -4,15 +4,15 @@ import lombok.Getter;
 
 @Getter
 public class Point {
-    int x;
-    int y;
+    double x;
+    double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int distanceTo(Point other) {
-        return (int) Math.sqrt(Math.pow(other.getX() - this.x, 2) + Math.pow(other.getY() - this.y, 2));
+    public double distanceTo(Point other) {
+        return Math.sqrt(Math.pow(other.getX() - this.x, 2) + Math.pow(other.getY() - this.y, 2));
     }
 }
